Initialize babylon

Executes bbi init ... in specified directory. This creates a babylon.yaml file, which contains defaults for many configurable babylon settings, in that directory.

bbi_init(
  .dir,
  .nonmem_dir,
  .nonmem_version = NULL,
  .no_default_version = FALSE
)

Arguments

.dir
Path to directory to run init in (and put the resulting babylon.yaml file)

.nonmem_dir
Path to directory with the NONMEM installation.

.nonmem_version
Character scalar for default version of NONMEM to use. If left NULL, function will exit and tell you which versions were found in .nonmem_dir

.no_default_version
If TRUE, force creation of babylon.yaml with no default NONMEM version. FALSE by default, and using TRUE is not encouraged.

Details

For rbabylon to make any calls out to bbi (for example in submit_model() or model_summary()) it must find a babylon.yaml file. The default behavior is to look for this file in the same directory as the model. submit_model() and submit_models() also support passing a configuration file via the .config_path argument.