Create a new Babylon server manager

The Babylon generator creates a new 'Babylon'-object, which contains logic to interact with a running bbq server from the babylon ecosystem.

create a new client for babylon queue

bbq_client(init = NULL, host = "http://localhost", port = 3333,
  verbose = TRUE, must_work = TRUE)

Arguments

init
whether to initialize bbq

host
hostname

port
port bbq listens on

verbose
more information about whats going on under the hood

must_work
must be able to ping a valid bbq server to initialize

Value

Babylon instance

Initialization

A new Babylon object is initialized with the new() method: bbq <- Babylon$new()

and can optionally take some initialization arguments:

init - whether to initialize a bbq instance on the local machine

host - the hostname the server is running on - defaults to 'http://localhost'

port - port the server is listening on - defaults to 3333.

verbose - whether internal status information should be displayed

mustWork - on initialization, check and confirm a bbq server is listening on the host/port configuration set

Methods

submit_models(modelpath, ...) - check ?submit_models for arguments, the server is automatically set internally

get_models(status, STATUSES) - get information about models that have been submitted

get_model(id) - get information about a model

poll(.ids, until, timeout, interval, print, parse) - poll a vector of models by ID name until completion

ids - vector if model ids

until - status criteria to poll against, default to COMPLETED or ERROR

timeout - length of time to poll before stopping

print - whether to print the status of all models each poll

parse - whether to parse the http response

Examples

# NOT RUN {
bbq <- Babylon$new()
bbq$get_models()

# get all ids
bbq$get_models %>% map_dbl("ID")

# find all queued models
bbq$get_models(status = "QUEUED") %>% map_dbl("ID")

bbq$poll(1:5) # poll for models 1-5 to complete

# get all directories for completed runs
bbq$get_models(status = "COMPLETED") %>% map_chr(~ .x$RunInfo$RunDir)
# }