Installs most current release of bbi

Identifies system running and pulls the relevant tarball for the current release of babylon from GitHub, and then installs it in the directory passed to .dir. If used in an interactive session, will open an installation menu confirming the installed version.

use_bbi(
  .dir = "/data/apps",
  .version = "latest",
  .force = FALSE,
  .quiet = FALSE
)

Arguments

.dir
directory to install bbi to on linux

.version
version of babylon to install. Must pass a character scaler corresponding to a tag found in https://github.com/metrumresearchgroup/babylon/releases

.force
If FALSE, the default, skips installation if requested version and local version are the same. If TRUE forces installation if it will be the same version.

.quiet
If TRUE, suppresses output printed to the console. FALSE by default.

Value

character