Summarize model outputs

Calls out to bbi and returns a named list of class bbi_{.model_type}_summary with model outputs and diagnostics.

model_summary(.mod, .bbi_args = NULL, ..., .dry_run = FALSE)

# S3 method for bbi_nonmem_model
model_summary(.mod, .bbi_args = NULL, ..., .dry_run = FALSE)

Arguments

.mod
Model to summarize.

.bbi_args
A named list specifying arguments to pass to babylon formatted like list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4). See print_bbi_args() for full list of options.

...
args passed through to bbi_exec()

.dry_run
show what the command would be without actually running it

Details

NONMEM

The returned list for a NONMEM model will contain the following top-level elements:

run_details

run_heuristics

parameters_data

parameter_names

ofv

condition_number

shrinkage_details

The summary call will error if it does not find certain files in the output folder. However, you can override this behavior with the following file-specific flags:

no_ext_file

no_grd_file

no_shk_file

If you are using an estimation method that does not produce any of the following files, or they are missing for some other legitimate reason, pass the appropriate flags through the .bbi_args argument. For example, if have asked to skip the $COV step, you would call model_summary(..., .bbi_args = list(no_cov_file = TRUE)).

Additionally, if you have renamed the .ext file from its default of <root>.ext you will need to pass ext_file = "NEWNAME" to .bbi_args.

Methods (by class)

bbi_nonmem_model: Get model summary from bbi_nonmem_model object