S3 generic for getting model summary

S3 generic for getting model summary

Get model summary from `bbi_nonmem_model` object and optionally check if model is done before erroring

S3 dispatch for getting model summary from output directory path

S3 dispatch for getting model summary from numeric input This will only work if you are calling from the same directory as the models, or if you have set the model directory with `set_model_directory()`

Run `bbi nonmem summary` and parse the output to a list

model_summary(
  .mod,
  .model_type = NULL,
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = NULL
)

# S3 method for bbi_nonmem_model
model_summary(
  .mod,
  .model_type = NULL,
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = NULL
)

# S3 method for character
model_summary(
  .mod,
  .model_type = c("nonmem", "stan"),
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = getOption("rbabylon.model_directory")
)

# S3 method for numeric
model_summary(
  .mod,
  .model_type = c("nonmem", "stan"),
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = getOption("rbabylon.model_directory")
)

nonmem_summary(.mod, .bbi_args = NULL, ..., .dry_run = FALSE)

Arguments

.mod
`bbi_nonmem_model` object for summary

.model_type
Character scaler specifying the type of model, either 'nonmem' or 'stan'

.bbi_args
A named list specifying arguments to pass to babylon formatted like `list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)`. Run `print_nonmem_args()` to see valid arguments.

...
args passed through to `bbi_exec()`

.dry_run
show what the command would be without actually running it

.directory
Model directory containing the files referenced by `.mod`. Defaults to `options('rbabylon.model_directory')`, which can be set globally with `set_model_directory()`.

Value

List of S3 Class "bbi_nonmem_summary" with all summary information